ncompress - (Ver. 4.2.3) an improved file compressor
This is version 4.2 of (N)compress (an improved version of compress 4.1).

Compress is a fast, simple LZW file compressor.  Compress does not have
the highest compression rate, but it is one of the fastest programs to
compress data.  Compress is the defacto standard in the UNIX community
for compressing files.

'build' is a menu driven shell script for compiling, testing and
installing (N)compress. So to build and install (N)compress all you have to
do is run build.  Build will first test your system for default
settings. The current compile settings are stored in a special file
called compress.def.

For user with problems with build there is a default makefile included
called 'Makefile.def'. Also build is capable te generate a Makefile with
all options (option genmake).

(N)compress 4.2 has a special, fast compression hash algorithm.  This
algorithm uses more memory than the old hash table. If you don't want
the faster hash table algorithm set 'Memory free for compress' below
800000.

zcmp, zdiff, zmore were copied from version 4.1 without any changes.

The output of (N)compress 4.2 is fully compatible with that of compress 3.0.
In other words, the output of (N)compress 4.2 may be fed into uncompress 3.0 or
the output of compress 3.0 may be fed into uncompress 4.2.

The output of (N)compress 4.2 is not compatible with that of
compress 2.0.  However, compress 4.2 still accepts the output of
compress 2.0.  To generate output that is compatible with compress
2.0, use the undocumented "-C" flag.