replace fixed stack size with filenames with dynamic buffers
This avoids the 1k path limit that MAXPATHLEN imposes.  Now we let the
OS throw an error if the paths are too long rather than always assuming
no path will be 1k+ bytes long.

It does mean we're allocating more heap memory than we were before, but
it shouldn't matter on any system built in the last two decades.