kick off new 5.x version series
The compress releases have had a tumultuous history.  The last official
version was 4.0 in Aug 1985.  All releases since have been intermittent
and completely unofficial.  In fact, ncompress itself is an unofficial
follow up to the unofficial 4.1 patches by Dave Mack.

The 4.2.4.x fork by myself was to make it clear that it was a follow up
to Peter's 4.2.4 release as I didn't have any history/context.  In case
Peter did another release, I didn't have to have any conflicts or there
to be confusion as to what's newer (he'd just release 4.2.5 or 4.3).

After quite a lot of digging, it's clear Peter's 4.2.x series was both
itself an unofficial update to compress, and a one-off from 1992.  As
my fork has been around long enough at this point (~15 years), and the
wider community have opted to switch to mine over anything else (*),
update the version series to reflect this by moving to 5.0.  No one
else is maintaining compress or ncompress projects, and the original
authors aren't around to bless any forks.

It's better to keep the ncompress name rather than reverting to the
old "compress" if only for clarity when searching Google.  Attempting
to search for "compress" will result in a ton of noise that cannot be
reasonably filtered out.  On the other hand, searching for "ncompress"
only finds relevant results :).

(*) Most BSD's have their own independent forks that date back to the
last official compress-4.0 release.  Since they maintain a single code
base with `compress` in it, it's unlikely they'd use any external code
regardless of licensing.